+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 -   [💎] Renaming `Custom Marker Icon URL` to `Custom marker URL` to follow Obsidian guidelines
+-   [💎] Showing a toast notice if the user uses an invalid combination of maki icon and custom marker url
 
 ## 1.1.0